(fix: example) Further code cleanup, moved all mopro related functions to `mopro` module.
